Move the pointer tests into a test file
